refactor(solver): split visitor content predicates
