+++ crates/tsz-solver/src/tests/file_size_baselines.txt
@@ -32,7 +32,7 @@ solver_file_subtype_generics 2217
 solver_file_call_args 2097
 solver_file_eval_index_access 2226
 solver_file_def_core 2244
-solver_file_visitor_predicates 2162
+solver_file_visitor_predicates 742
 solver_file_types 2018
 
 binder_max_loc        2890
